Fix naming conflicts in generate code > * update macro code > * move Ownable2Step module to ownable.rs
