+++ modules/src/lib.rs
@@ -11,6 +11,5 @@ pub mod erc20;
 pub mod erc721;
 pub mod erc721_receiver;
 pub mod erc721_token;
-pub mod ownable_2step;
 pub mod security;
 pub mod wrapped_native;
