+++ odra-macros/src/ast/deployer_item.rs
@@ -91,19 +91,19 @@ mod deployer_impl {
                     let caller = odra::EntryPointsCaller::new(env.clone(), |contract_env, call_def| {
                         match call_def.method() {
                             "init" => {
-                                let result = execute_init(contract_env);
+                                let result = __erc20_exec_parts::execute_init(contract_env);
                                 odra::ToBytes::to_bytes(&result).map(Into::into).unwrap()
                             }
                             "total_supply" => {
-                                let result = execute_total_supply(contract_env);
+                                let result = __erc20_exec_parts::execute_total_supply(contract_env);
                                 odra::ToBytes::to_bytes(&result).map(Into::into).unwrap()
                             }
                             "pay_to_mint" => {
-                                let result = execute_pay_to_mint(contract_env);
+                                let result = __erc20_exec_parts::execute_pay_to_mint(contract_env);
                                 odra::ToBytes::to_bytes(&result).map(Into::into).unwrap()
                             }
                             "approve" => {
-                                let result = execute_approve(contract_env);
+                                let result = __erc20_exec_parts::execute_approve(contract_env);
                                 odra::ToBytes::to_bytes(&result).map(Into::into).unwrap()
                             }
                             _ => panic!("Unknown method")
@@ -142,11 +142,11 @@ mod deployer_impl {
                     let caller = odra::EntryPointsCaller::new(env.clone(), |contract_env, call_def| {
                         match call_def.method() {
                             "total_supply" => {
-                                let result = execute_total_supply(contract_env);
+                                let result = __erc20_exec_parts::execute_total_supply(contract_env);
                                 odra::ToBytes::to_bytes(&result).map(Into::into).unwrap()
                             }
                             "pay_to_mint" => {
-                                let result = execute_pay_to_mint(contract_env);
+                                let result = __erc20_exec_parts::execute_pay_to_mint(contract_env);
                                 odra::ToBytes::to_bytes(&result).map(Into::into).unwrap()
                             }
                             _ => panic!("Unknown method")
